improved complexity of cloneDeep and cloneDeepFully
