+++ test/test.js
@@ -8,7 +8,7 @@ import { describe, mock, test } from "node:test";
 import { clearPolyfills, polyfill } from "./polyfills.js";
 
 import  cloneDeep from "../src/clone-deep/clone-deep.js";
-import cloneDeepFully from "../src/clone-deep-fully.js";
+import cloneDeepFully from "../src/clone-deep-fully/clone-deep-fully.js";
 import useCustomizers from "../src/use-customizers.js";
 
 import { CLONE, Tag } from "../src/utils/constants.js";
